starting to look at function hijack example

+++ src/cav.c
@@ -2,7 +2,7 @@
 #include "cav.h"
 #include <dlfcn.h>
 
-typedef int (*orig_SSL_get_verify_result_f_type)(const SSL *ssl);
+typedef long (*orig_SSL_get_verify_result_f_type)(const SSL *ssl);
 typedef int (*orig_do_handshake_f_type)(SSL *s);
 typedef int (*orig_SSL_connect_f_type)(SSL *s);
